requirements.txt
Django~=4.2.5
djangorestframework~=3.14.0
Pillow~=10.0.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
pep8-naming==0.13.2
django-debug-toolbar==3.4.0
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0